+++ examples/antd.js
@@ -5,8 +5,9 @@ import Menu, {SubMenu, Item as MenuItem, ItemGroup as MenuItemGroup, Divider} fr
 
 import 'rc-menu/assets/index.less';
 
-function handleSelect(selectedKey) {
-  console.log('selected ' + selectedKey);
+function handleSelect(info) {
+  console.log(info);
+  console.log('selected ' + info.key);
 }
 
 var titleRight = <span>sub menu</span>;
@@ -29,9 +30,7 @@ function render(container) {
       <SubMenu title={titleRight1} key="4">
         <MenuItem key="4-1">inner inner</MenuItem>
         <Menu.Divider/>
-        <SubMenu
-          openOnHover={false}
-          key="4-2"
+        <SubMenu key="4-2"
           title={titleRight2}
           >
           <MenuItem key="4-2-1">inn</MenuItem>
@@ -59,9 +58,7 @@ function render(container) {
       <SubMenu title={titleRight1} key="4">
         <MenuItem key="4-1">inner inner</MenuItem>
         <Menu.Divider/>
-        <SubMenu
-          openOnHover={false}
-          key="4-2"
+        <SubMenu  key="4-2"
           title={titleRight2}
           >
           <MenuItem key="4-2-1">inn</MenuItem>
@@ -80,8 +77,8 @@ function render(container) {
 
   var inlineMenu = (
     <Menu onSelect={handleSelect} mode="inline"
-          closeOnDeActive={false}
-          openOnHover={false}>
+          closeSubMenuOnDeactive={false}
+          openSubMenuOnHover={false}>
       <SubMenu title={titleRight} key="1">
         <MenuItem key="1-1">0-1</MenuItem>
         <MenuItem key="1-2">0-2</MenuItem>