+++ examples/single.js
@@ -6,12 +6,12 @@ import Menu, {SubMenu, Item as MenuItem, ItemGroup as MenuItemGroup, Divider} fr
 import 'rc-menu/assets/index.less';
 import 'font-awesome/css/font-awesome.css';
 
-function handleSelect(selectedKey) {
-  console.log('selected ' + selectedKey);
+function handleSelect(info) {
+  console.log('selected ', info);
 }
 
-function handleClick(selectedKey) {
-  console.log('click ' + selectedKey);
+function handleClick(info) {
+  console.log('click ', info);
 }
 
 var titleRight = <span>sub menu
@@ -33,7 +33,7 @@ render(container);
 function render(container) {
   var leftMenu = (
     <Menu onSelect={handleSelect}
-      onClick={handleClick}>
+          onClick={handleClick}>
       <SubMenu title={titleRight} key="1">
         <MenuItem key="1-1">0-1</MenuItem>
         <MenuItem key="1-2">0-2</MenuItem>
@@ -45,11 +45,8 @@ function render(container) {
       <SubMenu title={titleRight1} key="4">
         <MenuItem key="4-1">inner inner</MenuItem>
         <Menu.Divider />
-        <SubMenu
-          openOnHover={false}
-          key="4-2"
-          title={titleRight2}
-        >
+        <SubMenu key="4-2"
+                 title={titleRight2}>
           <MenuItem key="4-2-1">inn</MenuItem>
           <SubMenu title={titleRight3} key="4-2-2">
             <Menu>
@@ -65,6 +62,7 @@ function render(container) {
   );
   React.render(<div>
     <h2>single selectable menu</h2>
+
     <p>
       <button onClick={destroy}>destroy</button>
     </p>